Store your ACLED access information into your session.

Simple function to authenticate and store (through Sys.setenv()) ACLED access key for the acled_api() function. If email and key is stored via acled_access, the email and key arguments for acled_api can be NULL.

This function is deprecated and will stop functioning in September 2025. Please use your ACLED email and password to authenticate your requests in the acled_api() and associated functions which manage ACLED OAuth tokens for you.

Usage

acled_access(email, key)

Arguments

email

This is the email that you register in the ACLED Access portal (https://developer.acleddata.com/)

key

This is the key generated by the ACLED Access portal.

Value

Returns a success message if ACLED credentials are successfully authorized

See also

Examples

if (FALSE) { # \dontrun{
acled_access(email = "your_email", key = "your_key")
Sys.getenv("acled_email")
Sys.getenv("acled_key")
} # }