Request data from ACLED API

This function allows users to easily request data from the ACLED API. Users can include variables such as country, regions, dates of interest and the format (monadic or dyadic). The function returns a tibble of the desired ACLED events.

Usage

acled_api(
  email = NULL,
  password = NULL,
  country = NULL,
  regions = NULL,
  start_date = floor_date(Sys.Date(), "year") - years(1),
  end_date = Sys.Date(),
  timestamp = NULL,
  event_types = NULL,
  population = "none",
  inter_numeric = FALSE,
  monadic = FALSE,
  ...
)

Arguments

email

character string. Email associated with your ACLED account registered at https://acleddata.com/.

password

character string. The password associated with your ACLED account. If NULL, you will be prompted to enter your password interactively.

country

character vector. Default is NULL, which will return events for all countries. Pass a vector of countries names to retrieve events from specific countries. The list of ACLED countries. names may be found via acledR::acled_countries.

regions

vector of region names (character) or region codes (numeric). Default is NULL, which will return events for all regions. Pass a vector of regions names or codes to retrieve events from countries. within specific regions. The list of ACLED regions may be found via acledR::acled_regions.

start_date

character string. Format 'yyyy-mm-dd'. The earliest date for which to return events. The default is 1997-01-01, which is the earliest date available.

end_date

character string. Format 'yyyy-mm-dd'. The latest date for which to return events. The default is Sys.Date(), which is the most present date.

timestamp

numerical or character string. Provide a date or datetime written as either a character string of yyyy-mm-dd or as a numeric Unix timestamp to access all events added or updated after that date.

event_types

vector of one or more event types (character). Default is NULL, which will return data for all event types. To reurn data for only specific event types, request one or more of the following options (not case sensitive): Battles, Violence against civilians, Protests, Riots, Strategic Developments, and Explosions/Remote violence.

population

character. Specify whether to return population estimates for each event. It accepts three options: "none" (default), "best", and "full".

inter_numeric

logical. If FALSE (default), interaction code columns (inter1, inter2, and interaction) returned as strings describing the actor types/interactions. If TRUE, the values are returned as numeric values.

monadic

logical. If FALSE (default), returns dyadic data. If TRUE, returns monadic actor1 data.

...

string. Any additional parameters that users would like to add to their API calls (e.g. interaction or ISO)

Value

Returns a tibble of of ACLED events.

See also

Examples

if (FALSE) { # \dontrun{

# Get all the events coded by ACLED in Argentina from 01/01/2022 until 02/01/2022
# in dyadic-wide form
argen_acled <- acled_api(
   email = "youremail@mail.com", password = "password",
  country = "Argentina", start_date = "2022-01-01", end_date = "2022-02-01"
)

# tibble with all the events from Argentina where each row is one event.
argen_acled

# Get all events coded by ACLED in the Caribbean from 01/01/2022 to 10/01/2022
# in monadic-long form using email and key saved in environment

carib_acled <- acled_api(
  email = "youremail@mail.com", password = "password",
  regions = "Caribbean", start_date = "2022-01-01",
  end_date = "2022-01-10", monadic = TRUE,
)

## Tibble with all the events from the Caribbean where each row is one actor
carib_acled
} # }