Second dummy data frame of ACLED events emulating an old format, used in acled_deletion_api Vignette

Large dataset of multiple regions and countries, purposefully including deleted/modified events.

Usage

acled_old_deletion_dummy

Format

A data frame:

event_id_cnty

An unique individual identifier by number and country acronym (updated annually)

event_date

The day, month and year on which an event took place

year

The year in which an event took place

time_precision

A numeric code indicating the level of certainty of the date coded for the event

disorder_type

Type of disorder associated with the event and sub event type

event_type

The type of event

sub_event_type

The type of sub-event

actor1

The named actor involved in the event. Note: Actor 1 and Actor 2 do not imply directionality (e.g. attacker or defender)

assoc_actor_1

The named actor associated with or identifying actor1

inter1

A numeric code indicating the type of actor1

actor2

The named actor involved in the event. Note: Actor 1 and Actor 2 do not imply directionality (e.g. attacker or defender)

assoc_actor_2

The named actor associated with or identifying actor1

inter2

A numeric code indicating the type of actor1

interaction

A numeric code indicating the interaction between types of actor1 and actor2

civilian_targeting

Column referencing the presence of civilian targeting

iso

A numeric code for each individual country

region

The region of the world where the event took place

country

The country in which the event took place

admin1

The largest sub-national administrative region in which the event took place

admin2

The second largest sub-national administrative region in which the event took place

admin3

The third largest sub-national administrative region in which the event took place

location

The location in which the event took place

latitude

The latitude of the location

longitude

The longitude of the location

geo_precision

A numeric code indicating the level of certainty of the location coded for the event

source

The source of the event report

source_scale

The scale (local, regional, national, international) of the source

notes

A short description of the event

fatalities

The number of reported fatalities which occurred during the event

tags

Tags associated with the event.

timestamp

Numeric code of time

See also