Generate moving statistics

Generate moving statistics

Usage

generate_movers(
  data,
  var,
  unit_id,
  time_id,
  slide_funs,
  slide_periods,
  na.rm = T,
  complete = T
)

Arguments

data

ACLED data.

var

Variable of interest.

unit_id

Spatial unit of analysis (e.g., country, ADMIN1, region).

time_id

Temporal variable (e.g., week, month, year).

slide_funs

Requested moving statistics. Character vector with options including mean, median, sd, min, and max.

slide_periods

How many periods in the past to summarize over. Vector of one or more integers. Inf includes all previous periods.

na.rm

Whether to include NAs in the calculations.

complete

Whether to reqiure at least the requested time horizon to lapse prior to calculating moving statistics. For example, if slide_periods = 4, this would return NAs for the first 4 periods in the data and only begin calculating the moving statistic in the fifth time period. If FALSE, moving statistics for all available time periods are calculated, even if the requested time horizon has not elapsed.

Value

Returns a tibble grouped by unit_id.

See also

Examples

if (FALSE) {

# Request all events in India since 2018
df_india <- acled_api(countries = "India",
                      start_date = "2018-01-01",
                      end_date = "2022-08-30",
                      monadic = F)

# Aggregate to event counts per month across India
df_india_agg <-
   df_india %>%
   generate_counts(.,
   unit_id = "country",
   time_id = "event_date",
   time_target = "month")

# Generate 3 moving average of total events per month
df_india_agg_movers <-
   generate_movers(data = df_india_agg,
   var = "total_events",
   unit_id = "country",
   time_id = "event_month",
   slide_funs = "mean",
   slide_periods = 3)
}