Request data from ACLED API

This function allows users to easily request data in ACLED API. Users can include variables such as countries, regions, dates of interest and the type of file (monadic or dyadic). The function returns a tibble of the desired ACLED events.

Usage

acled_api(
  email = NULL,
  key = NULL,
  countries = NULL,
  regions = NULL,
  start_date = NULL,
  end_date = NULL,
  timestamp = NULL,
  event_types = NULL,
  monadic = FALSE,
  ...,
  acled_access = TRUE
)

Arguments

email

character string. Email associated with your ACLED account registered at https://developer.acleddata.com.

key

character string. Access key associated with your ACLED account registered at https://developer.acleddata.com.

countries

character vector. Default is NULL, which will return events for all countries. Pass a vector of country names to retrieve events from specific countries. The list of ACLED country names may be found via acledR::acled_countries,

regions

vector of region names (character) or region codes (numeric). Default is NULL, which will return events for all regions. Pass a vector of regions names or codes to retrieve events from countries within specific regions. The list of ACLED regions may be found via acledR::acled_regions,

start_date

character string. Format 'yyyy-mm-dd'. The earliest date for which to return events. The default is NULL, which will return events from all available time periods. If 'start_date' is NULL, 'end_date' must also be NULL.

end_date

character string. Format 'yyyy-mm-dd'. The latest date for which to return events. The default is NULL, which will return events from all available time periods. If 'end_date' is NULL, 'start_date' must also be NULL.

timestamp

numerical or character string. Provide a date or datetime written as either a character string of yyyy-mm-dd or as a numeric Unix timestamp to access all events added or updated after that date.

event_types

vector of one or more event types (character). Default is NULL, which will return data for all event types. To reurn data for only specific event types, request one or more of the following options (not case sensitive): Battles, Violence against civilians, Protests, Riots, Strategic Developments, and Explosions/Remote violence.

monadic

logical. If FALSE (default), returns dyadic data. If TRUE, returns monadic actor1 data.

acled_access

logical. If TRUE, you have used the acled_access function and the email and key arguments are not required.

Value

Returns a tibble of of ACLED events.

See also

Examples

if (FALSE) {

## Get all the events coded by ACLED in Argentina from 01/01/2022 until 02/01/2022 in dyadic-wide form
argen_acled <- acled_api(jane.doe.email,jane.doe.key,countries = "Argentina",start_date = "2022-01-01",end_date="2022-02-01", acled_access = FALSE)

## tibble with all the events from Argentina where each row is one event.
argen_acled

## Get all events coded by ACLED in the Caribbean from 01/01/2022 to 10/01/2022 in monadic-long form
carib_acled <- acled_api(john.doe.email,john.doe.key,regions = "Caribbean",start_date = "2022-01-01",end_date="2022-01-10", monadic=TRUE, acled_access = FALSE)

## Tibble with all the events from the Caribbean where each row is one actor
carib_acled
}