Request data from the ACLED Deletions API

This function allows users to pull deleted ACLED event IDs from the Deletions API.

Usage

acled_deletions_api(
  email = NULL,
  key = NULL,
  date_deleted = NULL,
  acled_access = TRUE
)

Arguments

email

character string. Email associated with your ACLED account registered at https://developer.acleddata.com.

key

character string. Access key associated with your ACLED account registered at https://developer.acleddata.com.

date_deleted

character string. Format 'yyyy-mm-dd' or Unix timestamp. The query will return all deleted events including and after the requested date/timestamp.

acled_access

logical. If TRUE it means that you have utilized the acled_access function and there is no need for the email and key arguments.

Value

Returns a tibble of ACLED data with columns for event_id_cnty and deleted_timestamp.

See also

Examples

if (FALSE) {

# Request deleted ACLED events since January 1, 2022
acled_deletions_api(date_deleted = "2022-01-01", acled_acess = TRUE)

}